[feat] 실시간 주식 가격 전송 기능 추가: StockWebSocketController 구현 및 StockPriceDto, StockPrice 엔티티 수정, Redis 설정 개선

+++ BE/src/main/java/com/stockr/be/domain/stock/entity/StockPrice.java
@@ -8,11 +8,11 @@ import java.time.Instant;
 @Builder
 public class StockPrice {
     private final String ticker;
-    private final double price;
-    private final double volume;
+    private final long price;
+    private final long volume;
     private final Instant timestamp;
 
-    public static StockPrice of(String ticker, double price, double volume) {
+    public static StockPrice of(String ticker, long price, long volume) {
         return StockPrice.builder()
                 .ticker(ticker)
                 .price(price)
